refactor(web): drop seedNewTask and useCreateTaskAndNavigate; navigate-only
Route loader now handles task creation; navigating to a fresh id is sufficient.
